drop the final remaining use of the deprecated pkg_resources module
Declaring a namespace package has gone through a few revisions.
pkg_resources has a version that is heavily deprecated. pkgutil provides
a python2/python3 compatible version that is also compatible with native
python3 namespaces.

https://packaging.python.org/en/latest/guides/packaging-namespace-packages/

pkg_resources is very very deprecated and importing or using it results
in deprecation warnings. It's time to move off of it entirely.

Fixes: https://github.com/PyFilesystem/pyfilesystem2/issues/577